# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin OpenApi",
        "description": "this is a api documentation",
        "contact": {
            "name": "Emad shirzad",
            "email": "shemad2022@gmail.com"
        },
        "license": {
            "name": "GPL-v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/config": {
            "get": {
                "tags": [
                    "Configs"
                ],
                "summary": "listAllItem",
                "description": "list all Item",
                "operationId": "107d362ad26109dd3ada26c0f348fce1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "Current page number",
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "data": {
                                            "description": "List of item",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ConfigModel"
                                            }
                                        },
                                        "first_page_url": {
                                            "description": "First page URL",
                                            "type": "string",
                                            "format": "uri"
                                        },
                                        "from": {
                                            "description": "First item number in the current page",
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "last_page": {
                                            "description": "Last page number",
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "links": {
                                            "description": "Links",
                                            "type": "array",
                                            "items": {
                                                "oneOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Previous"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Links"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Next"
                                                    }
                                                ]
                                            }
                                        },
                                        "last_page_url": {
                                            "description": "Last page URL",
                                            "type": "string",
                                            "format": "uri"
                                        },
                                        "next_page_url": {
                                            "description": "Next page URL",
                                            "type": "string",
                                            "format": "uri"
                                        },
                                        "path": {
                                            "description": "Path",
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "description": "Items per page",
                                            "type": "integer",
                                            "format": "int32"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/config/tel": {
            "get": {
                "tags": [
                    "Configs"
                ],
                "summary": "getConfigFromTel (get from channels)",
                "description": "get Config From Tel (get from channels)",
                "operationId": "0f6f714a5a9113759c9c8a96b09ab343",
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/config/setChannel": {
            "get": {
                "tags": [
                    "Configs"
                ],
                "summary": "setChannel",
                "description": "setChannel",
                "operationId": "ea17bc0aeac9244545308fcc9ea8a595",
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessModel": {
                "title": "Success Model",
                "description": "Success Model",
                "properties": {
                    "message": {
                        "description": "Success message",
                        "type": "string",
                        "format": "",
                        "example": "my message"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "title": "Error Model",
                "description": "Error Model",
                "properties": {
                    "message": {
                        "description": "error message",
                        "type": "string",
                        "format": "",
                        "example": "my error message"
                    },
                    "errors": {
                        "description": "errors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "error1",
                            "error2"
                        ]
                    }
                },
                "type": "object"
            },
            "Previous": {
                "title": "Previous",
                "description": "Represents a link",
                "properties": {
                    "url": {
                        "description": "Link URL",
                        "type": "string",
                        "example": null
                    },
                    "label": {
                        "description": "Link label",
                        "type": "string",
                        "example": "&laquo; Previous"
                    },
                    "active": {
                        "description": "Indicates whether the link is active",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Links": {
                "title": "Links ",
                "description": "Represents an active link",
                "properties": {
                    "url": {
                        "description": "Link URL",
                        "type": "string",
                        "example": "http://your-url/api/category?page=0"
                    },
                    "label": {
                        "description": "Link label",
                        "type": "string",
                        "example": "1"
                    },
                    "active": {
                        "description": "Indicates whether the link is active",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Next": {
                "title": "Next",
                "description": "Represents an active link",
                "properties": {
                    "url": {
                        "description": "Link URL",
                        "type": "string",
                        "example": null
                    },
                    "label": {
                        "description": "Link label",
                        "type": "string",
                        "example": "Next &raquo;"
                    },
                    "active": {
                        "description": "Indicates whether the link is active",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ConfigModel": {
                "title": "Config Model",
                "description": "Represents a config",
                "properties": {
                    "id": {
                        "description": "config ID",
                        "type": "integer",
                        "format": "int32"
                    },
                    "url": {
                        "description": "url",
                        "type": "string"
                    },
                    "type": {
                        "description": "type",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "http",
                "description": "Login with username and password to get the authentication token <div>Example: Bearer token</div>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}